Correction of implementation errors:
The code was changed to return the label of the object and not the entire URI.
The code of the class StatementContextExtractor was changed to have subclasses according to the variable partToBeExtracted.
A visitor was developed in class ObjectExtractor to eliminate the use of ifs.
The name of the main class was changed to ContextExtractor and its comment.
.gitignore and .gitattributes were added, copied from the project DBpedia spotlight.
The extraction of the type of an object is unnecessary and was removed. It was developed to return if an object is a URI, a literal or anonymous.